# go/docs/swagger.yaml
basePath: /
...
host: localhost:8080
info:
  contact:
    email: 65070209@kmitl.ac.th
    name: API Support
    url: https://github.com/Project-DSView/backend
  description: This is the backend API for DSView application In Services GO
  termsOfService: http://swagger.io/terms/
  title: DSView API Go Service
  version: 1.0.0-alpha
paths:
...
  /api/auth/google:
    get:
      description: Generate Google OAuth2 login URL and return to client
      produces:
      - application/json
      responses:
        "200":
          description: Returns auth URL and state
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with Google OAuth2
      tags:
      - auth
  /api/auth/google/callback:
    get:
      description: Handle callback from Google OAuth2, validate state, exchange code
        for token, and authenticate user
      parameters:
      - description: OAuth2 state parameter for CSRF protection
        in: query
        name: state
        required: true
        type: string
      - description: Authorization code from Google
        in: query
        name: code
        type: string
      - description: OAuth error from Google
        in: query
        name: error
        type: string
      responses:
        "302":
          description: Redirect to frontend with token or error page
          schema:
            type: string
        "400":
          description: Bad request - missing or invalid parameters
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - authentication failed
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Google OAuth2 callback handler
      tags:
      - auth
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API key authentication for secure endpoints.
    in: header
    name: dsview-api-key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"